requirements.txt
python-binance
dotenv
streamlit